Fix compiler warning: implicit conversion loses integer precision
src/gsb_file.c:161:18: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
                                        long_name = strlen (name);
                                                  ~ ^~~~~~~~~~~~~

+++ src/gsb_file.c
@@ -156,7 +156,7 @@ static void gsb_file_remove_old_backup (const gchar *filename)
 					gchar *tmp_year;
 					gchar *tmp_month;
 					gchar *tmp_day;
-					gint long_name;
+					size_t long_name;
 
 					long_name = strlen (name);
 					str_date = g_strndup (old_filename+long_name+1, 8);
